Allow to recv and send file descriptors when using EpollDomainSocketChannel.
Motiviation:

When using domain sockets on linux it is supported to recv and send file descriptors. This can be used to pass around for example sockets.

Modifications:
- Add support for recv and send file descriptors when using EpollDomainSocketChannel.
- Allow to obtain the file descriptor for an Epoll*Channel so it can be send via domain sockets.

Result:
recv and send of file descriptors is supported now.

+++ transport-native-epoll/src/main/java/io/netty/channel/epoll/EpollEventLoop.java
@@ -135,7 +135,7 @@ final class EpollEventLoop extends SingleThreadEventLoop {
     void add(AbstractEpollChannel ch) {
         assert inEventLoop();
         int id = nextId();
-        Native.epollCtlAdd(epollFd, ch.fd, ch.flags, id);
+        Native.epollCtlAdd(epollFd, ch.fd().intValue(), ch.flags, id);
         ch.id = id;
         ids.put(id, ch);
     }
@@ -145,7 +145,7 @@ final class EpollEventLoop extends SingleThreadEventLoop {
      */
     void modify(AbstractEpollChannel ch) {
         assert inEventLoop();
-        Native.epollCtlMod(epollFd, ch.fd, ch.flags, ch.id);
+        Native.epollCtlMod(epollFd, ch.fd().intValue(), ch.flags, ch.id);
     }
 
     /**
@@ -156,7 +156,7 @@ final class EpollEventLoop extends SingleThreadEventLoop {
         if (ids.remove(ch.id) != null && ch.isOpen()) {
             // Remove the epoll. This is only needed if it's still open as otherwise it will be automatically
             // removed once the file-descriptor is closed.
-            Native.epollCtlDel(epollFd, ch.fd);
+            Native.epollCtlDel(epollFd, ch.fd().intValue());
         }
     }
